+++ R/ref-grid.R
@@ -547,9 +547,9 @@ ref_grid <- function(object, at, cov.reduce = mean, mult.names, mult.levs,
     else {
         id = plyr::id(data[, nms, drop = FALSE], drop = TRUE)
         uid = !duplicated(id)
-        key = do.call(paste, data[uid, nms, drop = FALSE])
+        key = do.call(paste, unname(data[uid, nms, drop = FALSE]))
         key = key[order(id[uid])]
-        tgt = do.call(paste, grid[, nms, drop = FALSE])
+        tgt = do.call(paste, unname(grid[, nms, drop = FALSE]))
         wgt = rep(0, nrow(grid))
         for (i in seq_along(key))
             wgt[tgt == key[i]] = sum(data[["(weights)"]][id==i])